+++ app/tests/components/about-me/AboutMeIcon.spec.ts
