+++ app/tests/components/social/SocialCard.spec.ts
