+++ app/tests/components/the-footer/TheFooter.spec.ts
