+++ app/tests/components/ui/UiTextArea.spec.ts
@@ -1,21 +1,20 @@
 import type { VueWrapper } from "@vue/test-utils"
-import type { InputTypeHTMLAttribute } from "vue"
 
 import { shallowMount } from "@vue/test-utils"
 import { afterEach, beforeEach, describe, expect, test } from "vitest"
 
-import UiInput from "~/components/ui/UiInput.vue"
+import UiTextArea from "~/components/ui/UiTextArea.vue"
 
-describe("Компонент UiInput", () => {
+describe("Компонент UiTextArea", () => {
   const defaultProps = {
     id: "test-field",
-  }
+  } as const
   const labelText = "Текст label"
 
   let wrapper: VueWrapper
 
   beforeEach(() => {
-    wrapper = shallowMount(UiInput, {
+    wrapper = shallowMount(UiTextArea, {
       props: defaultProps,
     })
   })
@@ -24,17 +23,18 @@ describe("Компонент UiInput", () => {
     wrapper.unmount()
   })
 
-  const input = () => wrapper.find("input")
+  const textarea = () => wrapper.find("textarea")
   const label = () => wrapper.find("label")
-  const inputAttributes = (attribute: string) => input().attributes(attribute)
+  const textareaAttributes = (attribute: string) =>
+    textarea().attributes(attribute)
 
   const setLabel = async () => await wrapper.setProps({ label: labelText })
 
   test("Выставление атрибутов `id` и `name` на поле ввода", () => {
     // Проверка атрибута `id`
-    expect(inputAttributes("id")).toBe(defaultProps.id)
+    expect(textareaAttributes("id")).toBe(defaultProps.id)
     // Проверка атрибута `name`
-    expect(inputAttributes("name")).toBe(defaultProps.id)
+    expect(textareaAttributes("name")).toBe(defaultProps.id)
   })
 
   describe("Параметр label", () => {
@@ -63,40 +63,40 @@ describe("Компонент UiInput", () => {
 
   describe("Параметр placeholder", () => {
     test("Отсутствие по умолчанию", () => {
-      expect(inputAttributes("placeholder")).toBeUndefined()
+      expect(textareaAttributes("placeholder")).toBeUndefined()
     })
 
     test("Выставление при передаче параметра", async () => {
       const placeholder = "Тестовый placeholder"
       await wrapper.setProps({ placeholder })
 
-      expect(inputAttributes("placeholder")).toBe(placeholder)
+      expect(textareaAttributes("placeholder")).toBe(placeholder)
     })
   })
 
   describe("Параметр required", () => {
     test("По умолчанию true", () => {
       // При преобразовании в HTML, true заменяется на пустую строку.
-      expect(inputAttributes("required")).toBe("")
+      expect(textareaAttributes("required")).toBe("")
     })
 
     test("Выставление при передаче параметра", async () => {
       await wrapper.setProps({ required: false })
 
-      expect(inputAttributes("required")).toBeUndefined()
+      expect(textareaAttributes("required")).toBeUndefined()
     })
   })
 
-  describe("Параметр type", () => {
-    test("По умолчанию text", () => {
-      expect(inputAttributes("type")).toBe("text")
+  describe("Параметр rows", () => {
+    test("По умолчанию 4", () => {
+      expect(textareaAttributes("rows")).toBe("4")
     })
 
     test("Выставление при передаче параметра", async () => {
-      const type: InputTypeHTMLAttribute = "email"
-      await wrapper.setProps({ type })
+      const rows = 3
+      await wrapper.setProps({ rows })
 
-      expect(inputAttributes("type")).toBe(type)
+      expect(textareaAttributes("rows")).toBe(rows.toString())
     })
   })
 })
