+++ app/tests/composables/useFirstInputFocus.spec.ts
